Raise when PS did not create a stock_availables
We observed a bug in PrestaShop where when we create
a product, sometimes PrestaShop does not a record in the
'stock_availables' table, then we get an empty list of ids.
Fail hard so we are noticed about this issue. Otherwise,
we would just not be aware that no quantity has been sent.